shared refactor
