+++ src/Store/reducers/rootReducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 
-import authReducer         from './authReducer'
+import authReducer         from '../../Auth/Reducer'
 import loadingReducer      from './loadingReducer'
 
 export default combineReducers({
