+++ src/shared/Navbar/Link.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
 import { useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { stopLoading } from '../../Store/actions/Loading';
-import styled from 'styled-components'
 import { JsxElement } from 'typescript';
+import React from 'react';
+import styled from 'styled-components'
 
 
 const LinkItem = styled(Link)`
@@ -29,8 +29,8 @@ type Props = {
   className: Object
 }
 
-export default ( props: Props ) => {
-  const {direction, children, className} = props;
+export default (props: Props) => {
+  const { direction, children, className } = props;
 
   const dispatch = useDispatch();
   return (
