+++ src/shared/Navbar/linkDirection.js
@@ -1,5 +1,5 @@
-import { useParams } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { useSelector }  from 'react-redux';
+import { useParams }    from 'react-router-dom';
 
 export default (text) => {
   const houseParam = useParams().houseId;
