+++ src/shared/SideDrawer/SideDrawer.js
@@ -1,8 +1,8 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { CSSTransition } from 'react-transition-group';
+import React              from 'react';
+import ReactDOM           from 'react-dom';
+import { CSSTransition }  from 'react-transition-group';
 
-import './SideDrawer.css';
+import                          './SideDrawer.css';
 
 const SideDrawer = (props) => {
   const content = (
