+++ src/shared/Spinner/Spinner.js
@@ -1,6 +1,6 @@
-import React from 'react';
+import React    from 'react'
 
-import styles from './Spinner.module.css';
+import styles   from './Spinner.module.css'
 
 const Spinner = () => {
   return <div className={styles.loader}>Loading...</div>;
