--- src/shared/UIElements/AddButton/AddButton.module.css
@@ -1,14 +1,0 @@
-.addBtn {
-  width: 80%;
-  background-color: rgb(34, 34, 201);
-  color: white;
-  border: none;
-  height: 2.4rem;
-  border-radius: 10px;
-  margin-bottom: 0.8rem;
-  margin-top: 0.4rem;
-}
-
-.addBtn:active {
-  background-color: rgb(100, 59, 250);
-}
