+++ src/shared/UIElements/Backdrop.js
@@ -1,7 +1,7 @@
-import React from 'react';
+import React    from 'react';
 import ReactDOM from 'react-dom';
 
-import styles from './Backdrop.module.css';
+import styles   from './Backdrop.module.css';
 
 const Backdrop = (props) => {
   return ReactDOM.createPortal(
