+++ src/shared/UIElements/ErrorModal.js
@@ -1,7 +1,8 @@
-import React from 'react';
+import React  from 'react';
 
-import Modal from './Modal';
 import Button from '../FormElements/Button';
+
+import Modal  from './Modal';
 
 const ErrorModal = props => {
   return (
